Refactor Lemming class to track previous direction
This commit refactors the Lemming class in order to track the previous direction of the lemming. A new attribute, previousDirection, has been added to the class. This attribute is updated whenever the lemming changes its direction. The purpose of this change is to allow the lemming to revert back to its previous direction if it encounters a wall while moving downwards. Additionally, the move() method has been modified to handle this logic.